test(replay): split session-replay-vars.test.ts by actual subject (#1460)
session-replay-vars.test.ts contained zero references to
buildReplayBuiltinVars (the module it was named after) — it actually held
tests of replay/vars.ts, replay/script.ts, and runReplayScriptFile. Redistribute
each test to where its subject actually lives:

- vars.ts unit tests move to new src/replay/__tests__/vars.test.ts
- script.ts env-directive/actionLines tests move into script.test.ts (none
  were duplicates of existing coverage there)
- runReplayScriptFile tests merge into a new sibling,
  session-replay-runtime-maestro.test.ts, since they depend on a
  vi.mock('../../../core/dispatch.ts') configuration incompatible with the
  one already in session-replay-runtime.test.ts
- one genuine duplicate (`--update` env-directive regression test) is
  deleted in favor of the equivalent already in session-replay-runtime.test.ts

No production code changed.
