+++ src/replay/__tests__/vars.test.ts
@@ -1,0 +1,216 @@
+import { test } from 'vitest';
+import assert from 'node:assert/strict';
+import { AppError } from '../../kernel/errors.ts';
+import {
+  buildReplayVarScope,
+  collectReplayShellEnv,
+  parseReplayCliEnvEntries,
+  resolveReplayAction,
+  resolveReplayString,
+} from '../vars.ts';
+import { parseReplayScriptDetailed, readReplayScriptMetadata } from '../script.ts';
+import type { SessionAction } from '../../contracts/session-action.ts';
+
+const LOC = { file: 'test.ad', line: 1 };
+
+test('resolveReplayString substitutes variables', () => {
+  const scope = buildReplayVarScope({ fileEnv: { APP: 'settings' } });
+  assert.equal(resolveReplayString('open ${APP}', scope, LOC), 'open settings');
+});
+
+test('resolveReplayString supports fallback with :-default', () => {
+  const scope = buildReplayVarScope({});
+  assert.equal(resolveReplayString('wait ${WAIT_SHORT:-500}', scope, LOC), 'wait 500');
+});
+
+test('resolveReplayString prefers scope value over fallback', () => {
+  const scope = buildReplayVarScope({ fileEnv: { WAIT_SHORT: '1000' } });
+  assert.equal(resolveReplayString('wait ${WAIT_SHORT:-500}', scope, LOC), 'wait 1000');
+});
+
+test('resolveReplayString fallback preserves embedded braces via escapes', () => {
+  const scope = buildReplayVarScope({});
+  assert.equal(resolveReplayString('x ${A:-one\\}two}', scope, LOC), 'x one}two');
+});
+
+test('resolveReplayString throws on unresolved variable with file:line', () => {
+  const scope = buildReplayVarScope({ fileEnv: { OTHER: 'x' } });
+  assert.throws(
+    () => resolveReplayString('open ${MISSING}', scope, { file: 'a.ad', line: 7 }),
+    (error: unknown) =>
+      error instanceof AppError &&
+      error.code === 'INVALID_ARGS' &&
+      /Unresolved variable \$\{MISSING\} at a\.ad:7/.test(error.message),
+  );
+});
+
+test('resolveReplayString is case-sensitive', () => {
+  const scope = buildReplayVarScope({ fileEnv: { APP: 'settings' } });
+  assert.throws(() => resolveReplayString('${app}', scope, LOC), AppError);
+});
+
+test('resolveReplayString substitutes multiple vars on one line', () => {
+  const scope = buildReplayVarScope({ fileEnv: { A: '1', B: '2' } });
+  assert.equal(resolveReplayString('${A}-${B}-${A}', scope, LOC), '1-2-1');
+});
+
+test('buildReplayVarScope precedence: cli > shell > file > builtin', () => {
+  const scope = buildReplayVarScope({
+    builtins: { K: 'builtin' },
+    fileEnv: { K: 'file' },
+    shellEnv: { K: 'shell' },
+    cliEnv: { K: 'cli' },
+  });
+  assert.equal(scope.values.K, 'cli');
+
+  const shellWinsOverFile = buildReplayVarScope({
+    fileEnv: { K: 'file' },
+    shellEnv: { K: 'shell' },
+  });
+  assert.equal(shellWinsOverFile.values.K, 'shell');
+});
+
+test('collectReplayShellEnv strips AD_VAR_ prefix and ignores other vars', () => {
+  const result = collectReplayShellEnv({
+    AD_VAR_APP_ID: 'settings',
+    PATH: '/bin',
+    AD_VAR_123: 'x',
+    AD_VAR_: 'empty',
+    OTHER_VAR: 'y',
+    AD_APP_ID: 'no-legacy-prefix',
+  });
+  assert.equal(result.APP_ID, 'settings');
+  assert.equal(result.PATH, undefined);
+  assert.equal(result['123'], undefined);
+  assert.equal(result[''], undefined);
+  // legacy AD_* (non AD_VAR_*) is no longer auto-imported.
+  assert.equal(Object.prototype.hasOwnProperty.call(result, 'AD_APP_ID'), false);
+});
+
+test('collectReplayShellEnv skips keys that land in reserved AD_* namespace after strip', () => {
+  const result = collectReplayShellEnv({
+    AD_VAR_AD_SESSION: 'evil',
+    AD_VAR_AD_FOO: 'evil',
+  });
+  assert.equal(Object.prototype.hasOwnProperty.call(result, 'AD_SESSION'), false);
+  assert.equal(Object.prototype.hasOwnProperty.call(result, 'AD_FOO'), false);
+});
+
+test('parseReplayCliEnvEntries splits KEY=VALUE and rejects invalid keys', () => {
+  assert.deepEqual(parseReplayCliEnvEntries(['APP=settings', 'FOO=bar=baz']), {
+    APP: 'settings',
+    FOO: 'bar=baz',
+  });
+  assert.throws(() => parseReplayCliEnvEntries(['NOEQUAL']), AppError);
+  assert.throws(() => parseReplayCliEnvEntries(['lower=x']), AppError);
+  assert.throws(() => parseReplayCliEnvEntries(['=value']), AppError);
+});
+
+test('resolveReplayAction walks positionals and string flags', () => {
+  const action: SessionAction = {
+    ts: 0,
+    command: 'snapshot',
+    positionals: ['${FOO}'],
+    flags: {
+      snapshotScope: '${SCOPE}',
+      snapshotInteractiveOnly: true,
+      snapshotDepth: 3,
+    },
+  };
+  const scope = buildReplayVarScope({ fileEnv: { FOO: 'bar', SCOPE: 'app' } });
+  const resolved = resolveReplayAction(action, scope, LOC);
+  assert.deepEqual(resolved.positionals, ['bar']);
+  assert.equal(resolved.flags?.snapshotScope, 'app');
+  assert.equal(resolved.flags?.snapshotInteractiveOnly, true);
+  assert.equal(resolved.flags?.snapshotDepth, 3);
+});
+
+test('resolveReplayAction walks runtime hints', () => {
+  const action: SessionAction = {
+    ts: 0,
+    command: 'open',
+    positionals: [],
+    runtime: { platform: 'android', metroHost: '${HOST}' },
+    flags: {},
+  };
+  const scope = buildReplayVarScope({ fileEnv: { HOST: '10.0.0.1' } });
+  const resolved = resolveReplayAction(action, scope, LOC);
+  assert.equal(resolved.runtime?.metroHost, '10.0.0.1');
+});
+
+test('resolveReplayAction produces dispatch-ready literals for a realistic fixture', () => {
+  const script = [
+    'context platform=android',
+    'env APP_ID=settings',
+    'env WAIT_SHORT=500',
+    'env SETTINGS_ITEMS="label=Wait || label=Apps"',
+    '',
+    'open ${APP_ID} --relaunch',
+    'wait ${WAIT_SHORT}',
+    'click "${SETTINGS_ITEMS}"',
+    'is exists "${SETTINGS_ITEMS}"',
+    'snapshot -s "${SNAPSHOT_SCOPE:-app}"',
+  ].join('\n');
+  const metadata = readReplayScriptMetadata(script);
+  const parsed = parseReplayScriptDetailed(script);
+  const scope = buildReplayVarScope({
+    builtins: { AD_PLATFORM: 'android' },
+    fileEnv: metadata.env,
+    shellEnv: { APP_ID: 'shell-wins' },
+    cliEnv: { APP_ID: 'cli-wins' },
+  });
+  const resolved = parsed.actions.map((action, index) =>
+    resolveReplayAction(action, scope, {
+      file: 'fixture.ad',
+      line: parsed.actionLines[index] ?? 0,
+    }),
+  );
+  assert.deepEqual(resolved[0]?.positionals, ['cli-wins']);
+  assert.equal(resolved[0]?.flags.relaunch, true);
+  assert.deepEqual(resolved[1]?.positionals, ['500']);
+  assert.deepEqual(resolved[2]?.positionals, ['label=Wait || label=Apps']);
+  assert.deepEqual(resolved[3]?.positionals, ['exists', 'label=Wait || label=Apps']);
+  assert.equal(resolved[4]?.flags.snapshotScope, 'app');
+});
+
+test.each([
+  {
+    name: 'file env via parseReplayEnvLine',
+    run: () => readReplayScriptMetadata('context platform=android\nenv AD_FOO=bar\n'),
+    keyMatch: /AD_FOO/,
+  },
+  {
+    name: 'CLI -e via parseReplayCliEnvEntries',
+    run: () => parseReplayCliEnvEntries(['AD_FOO=x']),
+    keyMatch: /AD_FOO/,
+  },
+  {
+    name: 'buildReplayVarScope.fileEnv',
+    run: () => buildReplayVarScope({ fileEnv: { AD_FOO: 'x' } }),
+    keyMatch: /AD_FOO/,
+  },
+  {
+    name: 'buildReplayVarScope.cliEnv',
+    run: () => buildReplayVarScope({ cliEnv: { AD_SESSION: 'x' } }),
+    keyMatch: /AD_SESSION/,
+  },
+])('rejects AD_* as reserved namespace in $name', ({ run, keyMatch }) => {
+  assert.throws(
+    run,
+    (error: unknown) =>
+      error instanceof AppError &&
+      error.code === 'INVALID_ARGS' &&
+      /AD_\* namespace is reserved/.test(error.message) &&
+      keyMatch.test(error.message),
+  );
+});
+
+test('parseReplayCliEnvEntries error wording is user-friendly for invalid keys', () => {
+  assert.throws(
+    () => parseReplayCliEnvEntries(['lower=x']),
+    (error: unknown) =>
+      error instanceof AppError &&
+      error.code === 'INVALID_ARGS' &&
+      /uppercase letters, digits, and underscores/.test(error.message),
+  );
+});
